move parsing codes from `client` module to `parser` module
